+++ tests/functions/test_process_text.py
@@ -24,7 +24,7 @@ def test_column_name_type(process_test_df):
     with pytest.raises(TypeError):
         process_test_df.process_text(["text"])
 
-
+@pytest.mark.xfail(reason='new_column_names is deprecated.')
 def test_new_column_names_type(process_test_df):
     """Raise TypeError if `new_column_names` type is not string or list."""
     with pytest.raises(TypeError):
@@ -36,9 +36,9 @@ def test_new_column_names_type(process_test_df):
 def test_column_name_presence(process_test_df):
     """Raise ValueError if `column_name` is not in dataframe."""
     with pytest.raises(ValueError):
-        process_test_df.process_text(column_name="Test")
+        process_test_df.process_text(column_name="Test", string_function="lower")
 
-
+@pytest.mark.xfail(reason='new_column_names is deprecated.')
 def test_new_column_names_presence_str(test_returns_dataframe):
     """
     Raise ValueError if `new_column_names` is a str
@@ -52,7 +52,7 @@ def test_new_column_names_presence_str(test_returns_dataframe):
             pat=r"([ab])?(\d)",
         )
 
-
+@pytest.mark.xfail(reason='new_column_names is deprecated.')
 def test_new_column_names_presence_list(test_returns_dataframe):
     """
     Raise ValueError if `new_column_names` is a list and at least
@@ -66,7 +66,7 @@ def test_new_column_names_presence_list(test_returns_dataframe):
             pat=r"([ab])?(\d)",
         )
 
-
+@pytest.mark.xfail(reason='merge_frame is deprecated.')
 def test_merge_frame_type(test_returns_dataframe):
     """
     Raise TypeError if `merge_frame` type is not bool."""
@@ -79,7 +79,7 @@ def test_merge_frame_type(test_returns_dataframe):
             merge_frame="True",
         )
 
-
+@pytest.mark.xfail(reason='string_function must be present.')
 def test_string_function_is_None(process_test_df):
     """Test that dataframe is returned if string_function is None."""
     result = process_test_df.process_text(column_name="text")
@@ -99,7 +99,7 @@ def test_str_split(process_test_df):
 
     assert_frame_equal(result, expected)
 
-
+@pytest.mark.xfail(reason='new_column_names is deprecated.')
 def test_new_column_names(process_test_df):
     """
     Test that a new column name is created when
@@ -153,7 +153,7 @@ def test_str_cat_result_is_a_string(no_nulls_df):
 
     assert_frame_equal(result, expected)
 
-
+@pytest.mark.xfail(reason='new_column_names is deprecated.')
 def test_str_cat_result_is_a_string_and_new_column_names(no_nulls_df):
     """
     Test wrapper for Pandas ``.str.cat()`` method when the outcome is a string,
@@ -235,7 +235,7 @@ def returns_frame_1():
         }
     )
 
-
+@pytest.mark.xfail(reason='merge_frame is deprecated.')
 def test_return_dataframe_merge_is_None(returns_frame_1):
     """
     Test that the dataframe returned when `merge_frame` is None
@@ -249,7 +249,7 @@ def test_return_dataframe_merge_is_None(returns_frame_1):
     )
     assert_frame_equal(result, expected_output)
 
-
+@pytest.mark.xfail(reason='merge_frame is deprecated.')
 def test_return_dataframe_merge_is_not_None(returns_frame_1):
     """
     Test that the dataframe returned when `merge_frame` is not None
@@ -275,7 +275,7 @@ def test_return_dataframe_merge_is_not_None(returns_frame_1):
     )
     assert_frame_equal(result, expected_output)
 
-
+@pytest.mark.xfail(reason='merge_frame is deprecated.')
 def test_return_dataframe_merge_is_not_None_new_column_names_is_a_list(
     returns_frame_1,
 ):
@@ -305,7 +305,7 @@ def test_return_dataframe_merge_is_not_None_new_column_names_is_a_list(
     )
     assert_frame_equal(result, expected_output)
 
-
+@pytest.mark.xfail(reason='new_column_names is deprecated.')
 def test_return_dataframe_new_column_names_is_a_list_len_unequal(
     returns_frame_1,
 ):
@@ -329,20 +329,16 @@ def test_return_dataframe_new_column_names_is_a_list_len_unequal(
 
 def test_output_extractall(test_returns_dataframe):
     """
-    Test output when `string_function` is "extractall"
-    and `merge_frame` is None.
+    Raise ValueError if the output is a dataframe.
     """
-    expected_output = test_returns_dataframe["text"].str.extractall(
-        r"(?P<letter>[ab])?(?P<digit>\d)"
-    )
-    result = test_returns_dataframe.process_text(
+    with pytest.raises(ValueError):
+        test_returns_dataframe.process_text(
         column_name="text",
         string_function="extractall",
         pat=r"(?P<letter>[ab])?(?P<digit>\d)",
     )
-    assert_frame_equal(result, expected_output)
 
-
+@pytest.mark.xfail(reason='merge_frame is deprecated.')
 def test_output_extractall_merge_frame_is_not_None(test_returns_dataframe):
     """
     Test output when `string_function` is "extractall"